[FIX] academic: Fix issue when import partner_type for a bunch of records.
The mecanism of the constrain with a onchange is no correct for several resons, one of them
is affected in the import process with this fields set for a candidates records to create in parters.
The way to separate the part for backend in overide the write funcion solve this problem.

closes ingadhoc/odoo-academic#66

Signed-off-by: Augusto Weiss <awe@adhoc.com.ar>